Removed code out to a utils file to make export tool clean
